+++ lib/es_spacer/es_v_spacer.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 class EsVSpacer extends StatelessWidget{
 
   final bool big;
-  final int factor;
+  final double factor;
 
 
   EsVSpacer({this.big = false,this.factor=1});
